improve documentation, split code, update readme
