+++ examples/commands.py
@@ -1,3 +1,4 @@
+""" Example commands """
 import interactions
 from interactions import (
     slash_command,
@@ -20,7 +21,8 @@ from interactions import (
         )
     ],
 )
-async def test_slash(ctx: SlashContext, option: str = None) -> None:
+async def example_slash(ctx: SlashContext, option: str = None) -> None:
+    """ Example command """
     await ctx.defer(ephemeral=True)
 
     embed = Embed(
@@ -45,6 +47,7 @@ async def test_slash(ctx: SlashContext, option: str = None) -> None:
     description="List channel command",
 )
 async def list_channel_slash(ctx: SlashContext) -> None:
+    """ List channel command """
     await ctx.defer(ephemeral=True)
 
     channels = ctx.guild.channels
@@ -60,6 +63,7 @@ async def list_channel_slash(ctx: SlashContext) -> None:
     description="List role command",
 )
 async def list_role_slash(ctx: SlashContext) -> None:
+    """ List role command """
     await ctx.defer(ephemeral=True)
 
     roles = ctx.guild.roles
@@ -73,6 +77,7 @@ async def list_role_slash(ctx: SlashContext) -> None:
     description="List member command",
 )
 async def list_member_slash(ctx: SlashContext) -> None:
+    """ List member command """
     await ctx.defer(ephemeral=True)
 
     members = ctx.guild.members
@@ -84,6 +89,7 @@ async def list_member_slash(ctx: SlashContext) -> None:
 
 
 class MyExtension(interactions.Extension):
+    """ My extension class """
     @interactions.slash_command(
         name="ping",
         description="Ping command",
@@ -99,6 +105,7 @@ class MyExtension(interactions.Extension):
     async def ping_slash(
         self, ctx: interactions.SlashContext, option: str = None
     ) -> None:
+        """ Ping command """
         await ctx.defer(ephemeral=True)
 
         msg = await ctx.send("Please hold...")
